Fix build

+++ tests/quickstart_test/quickstart.sh
@@ -3,21 +3,33 @@
 set -x
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"	
 OPT_DIR=${DIR}/../../opt
+WD=$1
+: ${WD:=$DIR}
 
+echo running in $WD
+pushd $WD
 b37_source="ftp://ftp.1000genomes.ebi.ac.uk/vol1/ftp/technical/reference/phase2_reference_assembly_sequence/hs37d5.fa.gz"
 dbsnp_source="ftp://ftp.ncbi.nih.gov/snp/organisms/human_9606_b150_GRCh37p13/VCF/All_20170710.vcf.gz"
 
 ## Download reference and variant databases 
-wget $b37_source -O - | gunzip -c > hs37d5.fa
-wget http://web.stanford.edu/group/wonglab/varsim/insert_seq.txt
-wget http://web.stanford.edu/group/wonglab/varsim/GRCh37_hg19_supportingvariants_2013-07-23.txt
-wget $dbsnp_source -O All.vcf.gz
+if [[ ! -f hs37d5.fa ]];then
+    wget $b37_source -O - | gunzip -c > hs37d5.fa
+fi
+if [[ ! -f insert_seq.txt ]];then
+    wget http://web.stanford.edu/group/wonglab/varsim/insert_seq.txt
+fi
+if [[ ! -f GRCh37_hg19_supportingvariants_2013-07-23.txt ]];then
+    wget http://web.stanford.edu/group/wonglab/varsim/GRCh37_hg19_supportingvariants_2013-07-23.txt
+fi
+if [[ ! -f All.vcf.gz ]];then
+    wget $dbsnp_source -O All.vcf.gz
+fi
 
 ${OPT_DIR}/samtools-1.9_install/samtools faidx hs37d5.fa
 
 # Test run varsim to generate 1x coverage data
 export PATH=${OPT_DIR}/jdk1.8.0_131/bin:$PATH
-../../varsim.py --vc_in_vcf All.vcf.gz --sv_insert_seq insert_seq.txt \
+${DIR}/../../varsim.py --vc_in_vcf All.vcf.gz --sv_insert_seq insert_seq.txt \
 --sv_dgv GRCh37_hg19_supportingvariants_2013-07-23.txt \
 --reference hs37d5.fa --id simu --read_length 100 --vc_num_snp 3000000 --vc_num_ins 100000 \
 --vc_num_del 100000 --vc_num_mnp 50000 --vc_num_complex 50000 --sv_num_ins 2000 \
